Fixed DeepCode's errors
