+++ src/main/java/com/rednavis/vaadin/showcase/backend/service/password/CannotPerformOperationException.java
@@ -1,0 +1,12 @@
+package com.rednavis.vaadin.showcase.backend.service.password;
+
+public class CannotPerformOperationException extends Exception {
+
+  public CannotPerformOperationException(String message) {
+    super(message);
+  }
+
+  public CannotPerformOperationException(String message, Throwable source) {
+    super(message, source);
+  }
+}
