+++ src/main/java/com/rednavis/vaadin/showcase/backend/service/password/InvalidHashException.java
@@ -1,0 +1,12 @@
+package com.rednavis.vaadin.showcase.backend.service.password;
+
+public class InvalidHashException extends Exception {
+
+  public InvalidHashException(String message) {
+    super(message);
+  }
+
+  public InvalidHashException(String message, Throwable source) {
+    super(message, source);
+  }
+}
